refactor: 디렉토리 및 네이밍 수정
