+++ .eslintrc.cjs
@@ -6,7 +6,6 @@ module.exports = {
     'plugin:react/recommended',
     'plugin:react/jsx-runtime',
     'plugin:react-hooks/recommended',
-    // 'stylelint-config-standard',
     'prettier',
   ],
   ignorePatterns: ['dist', '.eslintrc.cjs'],
