+++ src/components/achievement/AchievementContainer.jsx
@@ -12,9 +12,6 @@ const AchievementContainer = () => {
       <AchievementCard />
       <AchievementCard />
       <AchievementCard />
-      <AchievementCard />
-      <AchievementCard />
-      <AchievementCard />
     </Flex>
   );
 };
