+++ src/components/achievement/index.js
@@ -1,0 +1,4 @@
+export { AchievementCard } from './AchievementCard';
+export { AchievementContainer } from './AchievementContainer';
+export { AchievementLabel } from './AchievementContainer';
+export { AchievementWrapper } from './AchievementContainer';
