+++ src/index.css
@@ -1,7 +1,7 @@
 @font-face {
   font-family: 'NanumGothic';
   font-weight: normal;
-  src: url('../fonts/NanumGothic-Regular.ttf');
+  src: url('../public/fonts/NanumGothic-Regular.ttf');
 }
 
 @font-face {
